refactor(indexing): separate SolrType and SolrField
For a future index schema validation, we need more
sophisticated definitions of Solr fieldTypes, field etc.

To depict that in code, the types enum is moved out
of SolrField and much extended.

The SolrField class is slightly extended to enable a
getName() shorthand to make deciding between getNameSearchable()
and getNameFacetable() unnecessary for most cases.