feat(index): add SolrCopyField to represent <copyField> #5989
A new class is introduced, including some more SolrFieldProperty
entries, used for <copyField>. (Also not used, might be nice for
validation, etc.)

As those fields are a equipped with a very small and static number of
properties, the constructors do not allow ingesting a map, but provide
all properties as constructor arguments.

<copyFields> link together other fields inside Solr, optionally using
dynamic behaviour via wildcard matching. We use an opinionated approach
here, only allowing for existing, defined fields from the schema to be
used (Solr allows for very arbitrary wildcard matches)