feat(index): add WIP of a schema model caching facade #5989
This EJB stores the Solr schema we need to follow as a single
source of truth. We will rely on this in-memory model to validate,
update and manage the schema inside the backing Solr instance.

It will be usable via API to reload the schema model from the database
(and code) plus will do so automatically during startup of Dataverse.
This is necessary to have empty Solr instances bootstrapped by us.

Thanks to @pkiraly @pdurbin and @rtreacy most of this code
was done during a "hacky friday" code-with-me session.