feat(api-test): add base class TestContainersBase and tag collection
- Adding edu.harvard.iq.dataverse.api.TestContainersBase as
  abstract class to serve as a base for API tests using Testcontainers
  backing.
- Also adding edu.harvard.iq.dataverse.api.util.Tags to have a
  collection of tags to use for JUnit5 tests (avoiding missspelling etc)

The Testcontainers usage is entirely optional and will only be triggered
when setting a system property `tc.enabled=true`. In this case,
test classes extending the base class can use the TC controlled
containers.

As we want to reuse containers between tests to boost performance
between the different API tests, the base class is deploying a custom
lifecycle control.