fix(ct-base): unpack Payara to target/payara
Payara 5 defaults to a "payara5" topmost dir, Payara 6 to
"payara6". To avoid adding different directories in the
assembly, cut the number from the directories name when
unpacking.

This does not prevent you from doing stupid things like
not cleaning before switching the version leading to an
unknown state of old and new libs, etc.